Unify HttpAsset and HttpAssetInterface
